Added files via upload
twitter blacklisting .Tweets are extracted using a web application and they are analysed using different data structures and search methods to filter sbusive and vulgar content. They are later sent to the NLP tool to determinde the postive negative and neutral tweets and the negative twets are dropped and postive ones are redirected to twitter.The neutral ones are left to the choice of the user under their own risk and circumstances.